insights clean up
