+++ .gitignore
@@ -113,3 +113,4 @@ website/static/js/*
 !website/static/js/mathjax.js
 !website/static/js/code_block_buttons.js
 website/static/_sphinx-sources/
+node_modules
