+++ captum/insights/frontend/src/App.module.css
@@ -113,6 +113,7 @@
 
 .btn {
   border: solid 1px #ee4c2c;
+  background: white;
   text-align: center;
   font-weight: 600;
   font-size: 1em;
