+++ captum/insights/frontend/src/components/Plot.js
@@ -1,0 +1,6 @@
+import createPlotlyComponent from "react-plotly.js/factory";
+import Plotly from "plotly.js-basic-dist";
+
+const Plot = createPlotlyComponent(Plotly);
+
+export default Plot;
