+++ captum/insights/frontend/src/components/Spinner.js
@@ -1,0 +1,8 @@
+import React from "react";
+import styles from "../App.module.css";
+
+function Spinner(_) {
+  return <div className={styles.spinner} />;
+}
+
+export default Spinner;
