+++ captum/insights/frontend/src/components/Tooltip.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+import styles from "../App.module.css";
+
+function Tooltip(props) {
+  return (
+    <div className={styles.tooltip}>
+      <div className={styles["tooltip__label"]}>{props.label}</div>
+    </div>
+  );
+}
+
+export default Tooltip;
